+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.0.9](https://github.com/mdvorak/esphome-adaptive-lighting/compare/v1.0.8...v1.0.9) (2025-12-17)
+
+
+### Bug Fixes
+
+* update AdaptiveLightingComponent for ESPHome 2025.12 compatibility in adaptive_lighting ([e28d36c](https://github.com/mdvorak/esphome-adaptive-lighting/commit/e28d36c1c231d4c9e5675f408b99d9f322304924))
+
+## [1.0.8](https://github.com/mdvorak/esphome-adaptive-lighting/compare/v1.0.7...v1.0.8) (2025-12-17)
+
+
+### Bug Fixes
+
+* ESPHome previous and new versions compatibility ([53da68f](https://github.com/mdvorak/esphome-adaptive-lighting/commit/53da68f07e23ef23d350f739e18fc1a08f836080))
+
+## [1.0.7](https://github.com/mdvorak/esphome-adaptive-lighting/compare/v1.0.6...v1.0.7) (2025-12-17)
+
+
+### Bug Fixes
+
+* migrate to ESPHome 2025.12 listener interface for light remote values callbacks ([c8d60bc](https://github.com/mdvorak/esphome-adaptive-lighting/commit/c8d60bce9346d35fcc11199ee0d21dfb38aedb4d))
+
 ## [1.0.6](https://github.com/mdvorak/esphome-adaptive-lighting/compare/v1.0.5...v1.0.6) (2025-07-03)